requirements.txt
requests
pytest
pandas
python-dotenv
